+++ rayna_engine/src/mesh/planar.rs
@@ -1,27 +1,29 @@
 //! This module is not an mesh module per-se, but a helper module that provides abstractions for
 //! planar types (such as planes, quads, triangles, etc)
 //!
-//! You should store an instance of [`Planar`] inside your mesh struct, and then simply validate the UV coordinates
+//! You should store an instance of [`Plane`] inside your mesh struct, and then simply validate the UV coordinates
 //! of the planar intersection for whichever shape your dreams do so desire...
 //!
 //! Most planar types ([`self::parallelogram::ParallelogramMesh`], [`self::infinite_plane::InfinitePlaneMesh`]) can't be instantiated directly,
-//! but can be easily converted via the [`From<Planar>`] conversion.
+//! but can be easily converted via the [`From<Plane>`] conversion.
 
 use crate::core::types::{Number, Point2, Point3, Vector3};
+use crate::mesh::Mesh;
+use crate::shared::aabb::{Aabb, Bounded};
 use crate::shared::intersect::Intersection;
 use crate::shared::interval::Interval;
 use crate::shared::ray::Ray;
 use getset::CopyGetters;
 use num_traits::Zero;
+use rand_core::RngCore;
 
-pub mod infinite_plane;
-pub mod parallelogram;
+// region Structs
 
 /// The recommended amount of padding around AABBs for planar objects
 ///
 /// Because planes are infinitely thin, we need to add padding to ensure they have at least some volume.
 /// Otherwise, there is a chance that the [`crate::shared::aabb::Aabb`] will always be missed because it has zero size.
-pub const AABB_PADDING: Number = 1e-6;
+pub const PLANAR_AABB_PADDING: Number = 1e-6;
 
 /// A helper struct that is used in planar objects (objects that exist in a subsection of a 2D plane)
 ///
@@ -29,7 +31,7 @@ pub const AABB_PADDING: Number = 1e-6;
 /// Then, you can restrict by validating the UV coordinates returned by the intersection
 #[derive(Copy, Clone, Debug, CopyGetters)]
 #[get_copy = "pub"]
-pub struct Planar {
+pub struct Plane {
     p: Point3,
     /// The vector for the `U` direction, typically the 'right' direction
     u: Vector3,
@@ -43,9 +45,77 @@ pub struct Planar {
     w: Vector3,
 }
 
+#[derive(Copy, Clone, Debug, CopyGetters)]
+#[get_copy = "pub"]
+pub struct ParallelogramMesh {
+    /// The plane that this mesh sits upon
+    plane: Plane,
+    aabb: Aabb,
+}
+
+#[derive(Copy, Clone, Debug, CopyGetters)]
+#[get_copy = "pub"]
+pub struct InfinitePlaneMesh {
+    /// The plane that this mesh sits upon
+    plane: Plane,
+    uv_wrap: UvWrappingMode,
+}
+
+// endregion Structs
+
+// region UV Wrap
+
+/// Enum for different ways UV coordinates can be wrapped (or not) on a plane
+#[derive(Copy, Clone, Debug, Default, Ord, PartialOrd, Eq, PartialEq)]
+pub enum UvWrappingMode {
+    // TODO: Remove `None`, add ones like clamp border, clamp edge
+    /// Wrap the UV coordinates when they reach `1.0`
+    ///
+    /// Equivalent to `x % 1.0`
+    #[default]
+    Wrap,
+    /// Mirror the UV coordinates when they reach `1.0`, repeating each interval
+    ///
+    /// Equivalent to `abs((x % 2.0) - 1.0)`
+    Mirror,
+    /// If either of the UV coordinates goes out of the range `0..=1`, sets both components to zero
+    ClampZero,
+    /// Clamps UV coordinates to the range `0..=1`
+    Clamp,
+}
+
+impl UvWrappingMode {
+    /// Applies the wrapping mode to the UV coordinate, returning the new coordinate
+    #[inline(always)]
+    pub fn apply(self, uvs: Point2) -> Point2 {
+        fn wrap(x: Number) -> Number { x.rem_euclid(1.0) }
+        fn mirror(x: Number) -> Number { ((x % 2.0) - 1.0).abs() }
+        fn clamp(x: Number) -> Number { x.clamp(0., 1.) }
+
+        match self {
+            Self::Wrap => Point2::new(wrap(uvs.x), wrap(uvs.y)),
+            Self::Mirror => Point2::new(mirror(uvs.x), mirror(uvs.y)),
+            Self::Clamp => Point2::new(clamp(uvs.x), clamp(uvs.y)),
+            Self::ClampZero => {
+                if !(0. <= uvs.x && uvs.x <= 1. && 0. <= uvs.y && uvs.y <= 1.0) {
+                    Point2::ZERO
+                } else {
+                    uvs
+                }
+            }
+        }
+    }
+
+    /// Applies the wrapping mode to the UV coordinate, writing the modified coordinate into the reference
+    #[inline(always)]
+    pub fn apply_mut(self, uvs: &mut Point2) { *uvs = self.apply(*uvs); }
+}
+
+// endregion UV Wrap
+
 // region Constructors
 
-impl Planar {
+impl Plane {
     /// Creates a plane from the origin point `p`, and the two side vectors `u`, `v`
     ///
     /// For a 2D plane in the `XY` plane, the point layout would be:
@@ -95,7 +165,7 @@ impl Planar {
         Self::new(centre - u - v, u * 2., v * 2.)
     }
 
-    /// Creates a [Planar] mesh from three points on the surface.
+    /// Creates a [Plane] mesh from three points on the surface.
     ///
     /// For a 2D plane in the `XY` plane, the point layout would be:
     ///
@@ -132,19 +202,51 @@ impl Planar {
 }
 
 /// Create from three point array
-impl<P: Into<Point3>> From<[P; 3]> for Planar {
+impl<P: Into<Point3>> From<[P; 3]> for Plane {
     fn from([p, a, b]: [P; 3]) -> Self { Self::new_points(p, a, b) }
 }
 /// Create from three point tuple
-impl<P: Into<Point3>, A: Into<Point3>, B: Into<Point3>> From<(P, A, B)> for Planar {
+impl<P: Into<Point3>, A: Into<Point3>, B: Into<Point3>> From<(P, A, B)> for Plane {
     fn from((p, a, b): (P, A, B)) -> Self { Self::new_points(p, a, b) }
+}
+
+impl ParallelogramMesh {
+    pub fn new(plane: impl Into<Plane>) -> Self {
+        let plane = plane.into();
+        let (p, a, b, ab) = (
+            plane.p(),
+            plane.p() + plane.u(),
+            plane.p() + plane.v(),
+            plane.p() + plane.u() + plane.v(),
+        );
+        let aabb = Aabb::encompass_points([p, a, b, ab]).with_min_padding(PLANAR_AABB_PADDING);
+
+        Self { plane, aabb }
+    }
+}
+
+impl InfinitePlaneMesh {
+    pub fn new(plane: impl Into<Plane>, uv_wrap: UvWrappingMode) -> Self {
+        Self {
+            plane: plane.into(),
+            uv_wrap,
+        }
+    }
+}
+
+impl<T: Into<Plane>> From<T> for ParallelogramMesh {
+    fn from(plane: T) -> Self { Self::new(plane) }
+}
+
+impl<T: Into<Plane>> From<T> for InfinitePlaneMesh {
+    fn from(plane: T) -> Self { Self::new(plane, UvWrappingMode::default()) }
 }
 
 // endregion
 
-// region Intersection
+// region Mesh Impl
 
-impl Planar {
+impl Plane {
     /// Does a full ray-plane intersection check, returning the intersection if possible. If an intersection is not found,
     /// it means that the ray is perfectly parallel to the plane, or outside the given interval.
     ///
@@ -193,4 +295,34 @@ impl Planar {
         })
     }
 }
-// endregion
+
+impl Mesh for InfinitePlaneMesh {
+    fn intersect(&self, ray: &Ray, interval: &Interval<Number>, _rng: &mut dyn RngCore) -> Option<Intersection> {
+        let mut i = self.plane.intersect_bounded(ray, interval)?;
+        // Wrap uv's if required
+        self.uv_wrap.apply_mut(&mut i.uv);
+        Some(i)
+    }
+}
+
+impl Mesh for ParallelogramMesh {
+    fn intersect(&self, ray: &Ray, interval: &Interval<Number>, _rng: &mut dyn RngCore) -> Option<Intersection> {
+        let i = self.plane.intersect_bounded(ray, interval)?;
+        // Check in interval for our segment of the plane: `uv in [0, 1]`
+        if (i.uv.cmple(Point2::ONE) & i.uv.cmpge(Point2::ZERO)).all() {
+            Some(i)
+        } else {
+            None
+        }
+    }
+}
+
+impl Bounded for InfinitePlaneMesh {
+    fn aabb(&self) -> Aabb { Aabb::INFINITE }
+}
+
+impl Bounded for ParallelogramMesh {
+    fn aabb(&self) -> Aabb { self.aabb }
+}
+
+// endregion Mesh Impl
