--- rayna_engine/src/mesh/primitive/mod.rs
@@ -1,4 +1,0 @@
-pub mod axis_box;
-pub mod cylinder;
-pub mod sphere;
-pub mod triangle;
