+++ rayna_engine/src/object/simple.rs
@@ -24,7 +24,7 @@ use rand_core::RngCore;
 /// Alternatively, you can also apply a post and pre-transform, to counteract the mesh's position offset:
 /// ```
 /// # use rayna_engine::material::lambertian::LambertianMaterial;
-/// # use rayna_engine::mesh::primitive::axis_box::AxisBoxMesh;
+/// # use rayna_engine::mesh::axis_box::AxisBoxMesh;
 /// # use rayna_engine::core::types::{Angle, Point3, Transform3, Vector3};
 /// #
 /// # let a: Point3 = [5., 1., 2.].into();
