+++ rayna_engine/src/scene/preset.rs
@@ -23,13 +23,13 @@ use crate::material::light::LightMaterial;
 use crate::material::metal::MetalMaterial;
 use crate::material::MaterialInstance;
 use crate::mesh::advanced::bvh_mesh::BvhMesh;
-use crate::mesh::isosurface::polygonised::PolygonisedIsosurfaceMesh;
-use crate::mesh::isosurface::raymarched::RaymarchedIsosurfaceMesh;
+use crate::mesh::axis_box::AxisBoxMesh;
 use crate::mesh::planar::infinite_plane::{InfinitePlaneMesh, UvWrappingMode};
 use crate::mesh::planar::parallelogram::ParallelogramMesh;
 use crate::mesh::planar::Planar;
-use crate::mesh::primitive::axis_box::AxisBoxMesh;
-use crate::mesh::primitive::sphere::SphereMesh;
+use crate::mesh::polygonised::PolygonisedIsosurfaceMesh;
+use crate::mesh::raymarched::RaymarchedIsosurfaceMesh;
+use crate::mesh::sphere::SphereMesh;
 use crate::mesh::MeshInstance;
 use crate::object::volumetric::VolumetricObject;
 use crate::object::ObjectInstance;
