+++ rayna_engine/src/shared/intersect.rs
@@ -37,7 +37,7 @@ pub struct Intersection {
     pub uv: Point2,
     /// Numeric ID for which "face" was hit
     ///
-    /// For objects with a single 'surface' (like a [sphere](crate::mesh::primitive::sphere::SphereMesh), this would be always zero.
+    /// For objects with a single 'surface' (like a [sphere](crate::mesh::sphere::SphereMesh), this would be always zero.
     /// For an mesh that may have multiple faces (like a [box](`crate::mesh::primitive::axis_box::AxisBoxMesh`), this would unique per-side.
     pub side: usize,
 }
